Added user-unique salted password hashing and user-transparent hash updating. If the first login fails, the client will automatically try sending a password change request to the server, then reattempt login after receiving a reply from the server. If it fails again, it will tell the user they have the wrong password.
Added checks for a username and password before trying to login when the player presses enter instead of clicking the Login button.

Currently password change requests are handled by the LoginManager. It should be clear enough where it is that it could be easily moved to its own Manager if desired.

Here are the packets.
Server->Client
ps = password change successful
pe = password change failed
Client->Server
cusername,newPassword,oldPassword