Draft done. Things left to do: - Get Evolution from Bulbapedia rather than old pokemon.ini - For habitat, not all pages have just a one word description... so what if we used the habitat sentence or two that each pokemon has on bulbapedia?
On my machine, with 1.5Mbit internet, this took 18 minutes to run.
We have to be careful with how often we run this, cause -- when I ran it too often, I was getting 503 errors... it could probably be solved by adding a couple seconds between the end of one pokemon and the start of the next. =/